requirements.txt
# Schema migrations
south

# For the admin documentation
docutils

# For the Imagefields
PIL

# For the API
django-piston